pyproject.toml
[project]
name = "snyk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=1.67.0",
    "pydantic-ai-slim[bedrock,logfire,mcp]>=1.67.0",
    "python-dotenv>=1.2.2",
]
